strchr & strrchr

+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: lukorman <lukorman@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/14 19:51:48 by lukorman          #+#    #+#             */
-/*   Updated: 2024/10/17 19:06:11 by lukorman         ###   ########.fr       */
+/*   Updated: 2024/10/21 15:01:38 by lukorman         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	*ft_memset(void *s, int c, size_t n)
 {
 	unsigned char	*ptr;
-	int				i;
+	size_t			i;
 
 	ptr = (unsigned char *)s;
 	i = 0;
